Move DocumentationPanel to components/common
