+++ airbyte-webapp/src/components/common/DocumentationPanel/DocumentationPanel.module.scss
@@ -1,4 +1,4 @@
-@use "../../scss/colors";
+@use "scss/colors";
 
 .container {
   min-height: 100vh;
