+++ airbyte-webapp/src/components/common/DocumentationPanel/DocumentationPanel.tsx
