+++ airbyte-webapp/src/components/common/DocumentationPanel/index.ts
