+++ airbyte-webapp/src/views/Connector/ConnectorDocumentationLayout/ConnectorDocumentationLayout.tsx
@@ -9,7 +9,7 @@ import styles from "./ConnectorDocumentationLayout.module.scss";
 import { useDocumentationPanelContext } from "./DocumentationPanelContext";
 
 const LazyDocumentationPanel = lazy(() =>
-  import("components/DocumentationPanel").then(({ DocumentationPanel }) => ({ default: DocumentationPanel }))
+  import("components/common/DocumentationPanel").then(({ DocumentationPanel }) => ({ default: DocumentationPanel }))
 );
 
 export const ConnectorDocumentationLayout: React.FC<React.PropsWithChildren<unknown>> = ({ children }) => {
